File Splitting for proxy and utils (#347)
* Uploading so I don't get too behind

* Uploading so I don't get too behind

* Black format and line fix

* Black format and line fix

* nox update

* nox update

* line fix

* Add files via upload

* GetRandomProxy to get_random_proxy
